move auth + settings + profile-edit routes to blueprints/auth.py
